set the version to 2.0-SNAPSHOT
The version 0.0.1-SNAPSHOT was used before upgrading to Spring Boot 2, and is still used by the Vaadin 8 version of the Patient Portal app. This change makes it easier to build both Vaadin 8 and Vaadin 14 versions: they depend on different versions of the backend module.

The `master` branch of this repo now builds the version 2.0-SNAPSHOT, which is Spring Boot 2.2 / Vaadin 14 compatible.
The version 0.0.1-SNAPSHOT,which is Spring Boot 1.4 / Vaadin 8 compatible, can be built from an earlier commit in this repo.